phase1245-band4-singletons-3-libs-to-100(platform+particle+imgui-backend)
All-modules-to-100 BAND 4 / singletons (60-65 -> 100). platform: FIXED 2 real compile defects
(X11Window.cpp + CocoaWindow.mm used non-existent enum names kArrowLeft/Right/Up/Down -> kLeft/
kRight/kUp/kDown; the Linux/macOS desktop backends would NOT compile) + desktop-first-v1 SEALED
(Cocoa/iOS/Android/Web/Wayland ports = per-platform promote-on-need) + 5 header-logic tests +
banner fix. particle_system: CPU-sim-v1 SEALED (GPU dispatch + force fields + max-particle-cap =
promote) + 6 real-branch tests (dt<=0 guard, fractional accumulator, large-dt multi-spawn, dead-
particle compaction, velocity bounds). imgui_backend: 'ZERO TESTS' state FIXED -> 9 host tests +
fixed a hidden pre-existing -Werror defect (ProfilerView row*row_height int*float); Vulkan-backend
-v1 SEALED (D3D12 ImGui + GPU-pixel verify = device-gated promote). Fixed 2 agent-introduced WAE
in the new imgui test (ctx_ member-init + auto-cast). ADR-20260616-band4-singletons-scope (3 §).
Build CLEAN -Werror; 5/5 PASS; golden BYTE-IDENTICAL. ==> Band 4 at 11/12 (editor last).

+++ engine/imgui_backend/include/cd/imgui/ProfilerView.hpp
@@ -109,7 +109,7 @@ inline void profiler_flamegraph(std::span<const cd::profile::Sample> samples,
                           / window_ns * static_cast<double>(canvas_w);
         const float fx0 = origin.x + static_cast<float>(x0);
         const float fx1 = origin.x + std::max(static_cast<float>(x1), static_cast<float>(x0) + 1.0F);
-        const float fy0 = origin.y + row * row_height;
+        const float fy0 = origin.y + static_cast<float>(row) * row_height;
         const float fy1 = fy0 + row_height - 2.0F;
         dl->AddRectFilled(ImVec2 { fx0, fy0 }, ImVec2 { fx1, fy1 },
                           detail::color_from_name_hash(s.name));
